backend/agents/qa_agent/requirements.txt
google-generativeai==0.3.2
flask==2.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
pytest==7.4.2
bandit==1.7.5
locust==2.20.0